Create and wire up new envoy receiver (#10)
This changeset creates a new open-telemetry-collector receiver as a new
Go module. It implements the envoy metricsserver and currently logs each
prometheus metric. The component is also included in `components.go` so
that it is included in the open-telemetry-collector. The component is
then included in the `inmem` `confprovider` so that it is included in
the running configuration.

The changeset also includes future plans for a set of golden tests for
the conversion from prometheus to otlp. The testdata was created by
running a envoy metricsserver instance and json encoding the envoy
metrics array to disk. The json library used for encoding was
`github.com/xhhuango/json` since the stdlib json encoded does not
support `NaN` values. We will need to use the same library or another
that support's `NaN` for reading in testing.

Future changesets will include metric conversion and pushing the data
onto the the pipeline.