Restructuring of the otel packages and configuration
A few changes were made here to try to organize our base configuration
passed to the otel sdk.

1. Removed the map of interfaces and situated around the map[string]component.IDs
2. Seperated the building of the Exporters/Receivers/Processsors/Extensions into
a default builder and then building up of the pipelines on each of these to try
and simplify the abstraction.
3. Moved packages around to better organize the packages.
   - otel/config : package handles all of the configuration building
   - otel/collector/providers : holds all the providers which consume the otel/config
   - otel/collector : the buidling of the resolver which uses the providers.

None of this is locked in stone and we can definitely change it but this gives us a first pass
at restructuring this work for continuing clarity.