Convert envoy metrics to otlp (#19)
Envoy metrics are emitted to MetricsServiceServer as prometheus protobuf
metrics. They are different from consuming a prometheus endpoint but
still a prometheus format. This means that we have to translate them
differently from the open-telemetry specification:
https://opentelemetry.io/docs/reference/specification/compatibility/prometheus_and_openmetrics/

They are different in the following ways:
* Envoy emits both histograms _and_ summaries by default. We're only
going to consume the histogram (which is also what ends up on the envoy
prometheus endpoint)
* Histograms are delta histograms rather than cumulative histograms