Include metric filter in hcp pipeline (#21)
Include the filter processor in the HCP metric pipeline. The metric
filters are pulled in from the `hcp.TelemetryClient`. A new function
called is added to insert new processors between the `MemoryLimiter` and
the `Batch` processor.

The `Filter` processor is added with a new `opt` function which inserts
the currently only supported filter processor. New changes would need to
be made to add new filter cases in the processor configuration and the
service config to support new named filters (if necessary).

The golden tests are also updated to reflect the new expected
configuration.