Initial e2e kubernetes workflow (#57)
Create a github action workflow that will run e2e tests against
consul-k8s. The workflow will perform a `workflow_dispatch` against the
hashicorp internal consul-k8s-workflow. This will happen when the e2e
label is attached to the PR.

Use the last 2 commits to observe and verify the behavior