Introduce new Exporter interface
The new Exporter interface allows the insertion of Configuration into
the external open-telemetry Pipeline. The Exporter interface is expected
to provide a *confmap.Conf which can be merged in with the rest of the
built configuration. Each implemention is responsible for ensuring that
it provides the correct configuration for that matching `component.ID`