PR #88 suggestion: remove the exporter interface (#91)
I suggested something in the PR, and felt like it might be easier to
show what I mean by making a small PR on top:

- Removes the `type Exporter interface` interface, instead extend the
`OtlpExporterCfg` method to perform the merge.
- In `config.go`, use the method directly : `cfg, err :=
exporters.OtlpExporterCfg(p.ExporterConfig.Exporter)`

The above also follows the pattern of what we do for the HCP provider,
so it's more consistent, i.e. :

https://github.com/hashicorp/consul-telemetry-collector/blob/8b03bae3d15cc7e4216295eecb7dbed5b1a2eaf6/internal/otel/config/config.go#L143

- Update tests to validates that configuration gets merged as expected

Additional small cleanups:
- Remove the unused `type OTLPExporterConfig struct`