Create more otlphttp configuration options (#88)
This changeset creates a new block of configuration for the otlphttp
exporter allowing the user more configuration options than before. The
previous `http_forwarder_endpoint` is now deprecated and warnings are
logged when both values are set. The configuration is then translated to
an internal type to generate the actual exporter configuration.

Unit testing was performed to ensure that we are still shipping the same
otel configuration to the otel server and new configuration options were
exercised in the test cases.

A future PR will handle pulling in the otelgrpc exporter

---------

Co-authored-by: Ashvitha <ashvitha.sridharan@hashicorp.com>